# packages/contracts/openapi.json
{
...
  "openapi": "3.1.0",
  "info": {
    "title": "Bookflow API",
    "version": "0.0.0",
    "description": "Generated from the Zod route schemas in apps/api. Never hand-written (ADR-014, ADR-025)."
  },
  "components": {
    "schemas": {
...
      "RenameBusinessRequestInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Required. Trimmed. 1–200 characters after trimming."
          },
          "tagline": {
            "type": "string",
            "maxLength": 200
          },
          "about": {
            "type": "string",
            "maxLength": 2000
          },
          "category": {
            "type": "string",
            "maxLength": 100
          },
          "address": {
            "type": "string",
            "maxLength": 500
          },
          "mapsUrl": {
            "type": "string",
            "maxLength": 2000,
            "format": "uri"
          }
        },
        "required": [
          "name"
        ],
        "description": "The business’s editable profile. The name is required."
      },
      "CreateBusinessRequestInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Required. Trimmed. 1–200 characters after trimming."
          }
        },
        "required": [
          "name"
        ],
        "description": "The business to create. Name only."
      },
...
      "Business": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "name": {
            "type": "string"
          },
          "published": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "published"
        ],
        "additionalProperties": false,
        "description": "A business the caller is a member of."
      },
...
    }
  },
  "paths": {
...
    "/v1/businesses/{businessId}": {
      "get": {
        "operationId": "getBusiness",
        "summary": "A business the caller belongs to",
        "tags": [
          "businesses"
        ],
        "description": "Scoped through membership: user → membership → business. A business the caller has no membership in is indistinguishable from one that does not exist.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "businessId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Business"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "renameBusiness",
        "summary": "Edit a business the caller belongs to",
        "tags": [
          "businesses"
        ],
        "description": "The name is required; tagline, about, category, address and mapsUrl are optional and an omitted one is left UNCHANGED, not cleared. Scoped through membership: a business the caller has no membership in is indistinguishable from one that does not exist. Text is trimmed before it is stored.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameBusinessRequestInput"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "businessId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Business"
                }
              }
            }
          }
        }
      }
    },
    "/v1/businesses": {
      "post": {
        "operationId": "createBusiness",
        "summary": "Create the caller's business",
        "tags": [
          "businesses"
        ],
        "description": "Creates the business and the caller's owner membership in one statement, so neither can exist without the other. An account may hold only one business (ADR-003): a second attempt is refused with 409 business-already-exists and writes nothing. The name is trimmed before it is stored.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusinessRequestInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Business"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/me/business/portfolio-images/{imageId}": {
      "delete": {
        "operationId": "deletePortfolioImage",
        "summary": "Remove a gallery image",
        "tags": [
          "media"
        ],
        "description": "Removes the row and then the stored object. If the object cannot be removed the request still succeeds — the image is off the page, which is what was asked, and the orphan is logged.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "imageId",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          }
        }
      }
    },
...
  }
}